restructure into package
