+++ runserver.py
@@ -1,0 +1,3 @@
+
+from ezproxylookup import app
+app.run(debug=True)
